Refactor header includes for cross-platform compatibility; add Texture class for image handling
