+++ src/Toolbar.h
@@ -1,7 +1,11 @@
 #ifndef TOOLBAR_H
 #define TOOLBAR_H
 
-#include "../bobcat_ui/all.h"
+#if defined(__APPLE__)
+    #include "../bobcat_ui/all.h"
+#else
+    #include <bobcat_ui/all.h>
+#endif
 #include "Enums.h"
 
 class Toolbar : public bobcat::Group {
