--- src/stb_image_impl.cpp
@@ -1,6 +1,0 @@
-//
-// Created by Hongzhe Xie on 5/3/25.
-//
-
-#define STB_IMAGE_IMPLEMENTATION
-#include "stb_image.h"
